Revert "feat: Export property groups"
This reverts commit e27175459441a46cb44babd82ab3256f9dc899a3.
