+++ src/index.js
@@ -1,4 +1,24 @@
-const propertyGroups = require('./groups')
+const { positioning } = require('./groups/positioning')
+const { layout } = require('./groups/layout')
+const { boxModel } = require('./groups/box-model')
+const { appearance } = require('./groups/appearance')
+const { typography } = require('./groups/typography')
+const { interaction } = require('./groups/interaction')
+const { transition } = require('./groups/transition')
+const { svgPresentation } = require('./groups/svg-presentation')
+
+const propertyGroups = [
+  ['composes'],
+  ['all'],
+  interaction,
+  positioning,
+  layout,
+  boxModel,
+  typography,
+  appearance,
+  svgPresentation,
+  transition
+]
 
 const propertiesOrder = propertyGroups.map((properties) => ({
   noEmptyLineBetween: true,
